llm-api/pyproject.toml
[project]
name = "llm-api"
version = "1.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
 "ai-shared",
 "fastapi>=0.123.5",
 "intime-common-packages",
 "pydantic-settings>=2.12.0",
 "python-multipart>=0.0.20",
 "pytz>=2025.2",
 "uvicorn>=0.38.0",
]

[tool.uv.sources]
ai-shared = { path = "dependencies/ai_shared-1.0.1-py3-none-any.whl" }
intime-common-packages = { path = "dependencies/intime_common_packages-0.4.13-py3-none-any.whl" }
